requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
pyPDF2
faiss-cpu
langchain_google_genai
langchain-community